Move the contents of the microlensing lightcurve label per time step database to a general lightcurve label per time step database
